requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
schedule==1.2.0
PyJWT==2.8.0
esptool==4.6.2

